Account for 2s resolution in exfat
By default, treat mtime checks with a difference of '1' as passing.
This is useful for exfat filesystem that has a reduced 2s resolution
for mtime and will produce false positives when checking for changes
in mtime. To switch this off, use `--strict-mtime'.

The default behavior should be fine for the vast majority of cases.